refactor(commands): extract attachment tests to sibling file
